requirements.txt
pandas
numpy
pyyaml
pydantic
pyodbc
matplotlib
wordcloud
nltk
scikit-learn
fastapi
uvicorn
textblob
-e .